+++ crates/multicalc/src/scalar/mod.rs
@@ -9,9 +9,11 @@ pub mod function;
 pub mod hyper_dual;
 pub mod jet;
 pub mod numeric;
+pub mod primal;
 
 pub use dual::Dual;
 pub use function::{ScalarFn, ScalarFnN, VectorFn, c};
 pub use hyper_dual::HyperDual;
 pub use jet::Jet;
 pub use numeric::Numeric;
+pub use primal::Primal;